requirements.txt
Flask
google-generativeai
protobuf 
PyPDF2
python-pptx
Pillow
requests
playwright
